7sprint 2

+++ src/scripts/card.js
@@ -1,5 +1,6 @@
 export const createCard = (
   cardData,
+  myId,
   onDeleteCard,
   onLikeCard,
   onOpenImagePopup,
@@ -9,9 +10,16 @@ export const createCard = (
   const cardImage = cardElement.querySelector(".card__image");
   cardImage.src = cardData.link;
   cardImage.alt = "Фото " + cardData.name;
+  const cardOwner = cardData.owner["_id"];
+  console.log(cardOwner, myId);
   cardElement.querySelector(".card__title").textContent = cardData.name;
+
   const deleteButton = cardElement.querySelector(".card__delete-button");
-  deleteButton.addEventListener("click", () => onDeleteCard(cardElement));
+  if (cardOwner === myId) {
+    deleteButton.addEventListener("click", () => onDeleteCard(cardElement));
+  }
+  else {deleteButton.style.display = 'none';}
+
   const likeButton = cardElement.querySelector(".card__like-button");
   likeButton.addEventListener("click", () => onLikeCard(likeButton));
   cardImage.addEventListener("click", () => onOpenImagePopup(cardData));
